BREAKING CHANGE: Update SimpleMath.kt

+++ feature-module/kotlin-lib/src/main/java/com/example/kotlin_lib/SimpleMath.kt
@@ -2,7 +2,7 @@ package com.example.kotlin_lib
 
 class SimpleMath {
 
-    fun sum(a: Int, b: Int): Int = a + b
+    fun sum(a: Long, b: Long): Long = a + b
 
     fun sub(a: Int, b: Int): Int = a - b
 }
